requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
pyjwt==2.10.1
streamlit==1.41.1
streamlit-autorefresh==1.0.1
pandas==2.2.3
chromadb==0.6.3
langchain==0.3.25
langchain-core==0.3.59
langchain-community==0.3.24
langchain-ollama==0.3.2
python-docx==1.1.2
pytest==8.3.5
ollama==0.4.7
lxml==5.3.0
